travio/requirements.txt
Fabric==1.10.2
Jinja2==2.8
Markdown==2.6.5
MarkupSafe==0.23
Pillow==3.0.0
PyJWT==1.4.0
PyYAML==3.10
Pygments==2.0.2
amqp==1.4.7
argparse==1.2.1
autopep8==1.2.1
billiard==3.3.0.21
celery==3.1.19
cloudinary==1.2.0
colorama==0.2.5
django-appconf==1.0.1
django-celery==3.1.17
django-compressor==1.5
django-debug-toolbar==1.4
django-extensions==1.5.9
django-filter==0.11.0
django-redis==4.3.0
django-redisboard==2.0.0
django-silk==0.5.6
django-simple-captcha==0.5.1
django-suit==0.2.15
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
ecdsa==0.13
html5lib==0.999
httpie==0.9.2
httplib2==0.9.2
jsonpatch==1.3
jsonpointer==1.0
kombu==3.0.29
newrelic==2.58.1.44
oauth==1.0.1
paramiko==1.16.0
pep8==1.6.2
prettytable==0.7.2
pyOpenSSL==0.13
pycrypto==2.6.1
pycurl==7.19.3
python-dateutil==2.4.2
python-firebase==1.2
pytz==2015.7
redis==2.10.5
requests==2.8.1
simplejson==3.8.1
six==1.5.2
ssh-import-id==3.21
uWSGI==2.0.11.2
urllib3==1.7.1
wheel==0.24.0
zope.interface==4.0.5
